requirements.txt
ccxt==4.1.13
flask==2.3.3
flask-cors==4.0.0
pandas>=1.5.0
numpy>=1.21.0
requests==2.31.0
python-dotenv==1.0.0
websocket-client==1.6.3
plotly==5.17.0
dash==2.14.1
dash-bootstrap-components==1.4.1 